# api/api.yml
openapi: 3.0.3
info:
  title: Product/Service Catalog API
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Local development server
paths:
  /products:
    get:
      summary: List all products
      operationId: listProducts
      parameters:
        - name: sort
          in: query
          description: Sort order
          required: false
          schema:
            type: string
        - name: filter
          in: query
          description: Filter criteria
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of products to return
          required: false
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          description: Number of products to skip before starting to collect the result set
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    post:
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /products/{id}:
    get:
      summary: Get a product by ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal server error
    put:
      summary: Update a product by ID
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a product by ID
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal server error
  /products/bulk:
    post:
      summary: Create multiple products
      operationId: bulkCreateProducts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Products created
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    put:
      summary: Update multiple products
      operationId: bulkUpdateProducts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Products updated
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    delete:
      summary: Delete multiple products
      operationId: bulkDeleteProducts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '204':
          description: Products deleted
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time